% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acPCAtuneLambda.R
\name{acPCAtuneLambda}
\alias{acPCAtuneLambda}
\title{Tune the lambda parameter in function acPCA}
\usage{
acPCAtuneLambda(
  X,
  Y,
  nPC,
  lambdas,
  centerX = T,
  centerY = T,
  scaleX = F,
  scaleY = F,
  kernel = c("linear", "gaussian"),
  bandwidth = NULL,
  anov = T,
  perc = 0.05,
  quiet = F
)
}
\arguments{
\item{X}{the n by p data matrix, where n is the number of samples, p is the number of variables. Missing values in X should be labeled as NA. If a whole sample in X is missing, it should be removed.}

\item{Y}{the n by q confounder matrix, where n is the number of samples, q is the number of confounding factors. Missing values in Y should be labeled as NA.}

\item{nPC}{number of principal components to compute.}

\item{lambdas}{a vector with the tuning parameters, non-negative values. If 0 is not in lambdas, it will be added to lambdas.}

\item{centerX}{center the columns in X. Default is True.}

\item{centerY}{center the columns in Y. Default is True.}

\item{scaleX}{scale the columns in X to unit standard deviation. Default is False.}

\item{scaleY}{scale the columns in Y to unit standard deviation. Default is False.}

\item{kernel}{the kernel to use: "linear", "gaussian".}

\item{bandwidth}{bandwidth h for Gaussian kernel. Optional.}

\item{anov}{True or False. Whether the penalty term has the between groups sum of squares interpretation. Default is True.}

\item{perc}{the best lambda is defined to be the smallest lambda with R(lambda)<=perc (if anov=T), or R(lambda)<=perc*R(lambda=0) (if anov=F) in the nPC principal components.}

\item{quiet}{True or False. Output the progress of the program. Default is False.}
}
\value{
Results for tuning lambda
\item{ratio}{R(lambda): a vector with the ratios. Same length as lambdas}
\item{best_lambda}{the best lambda after cross-validation}
\item{...}{Input parameters for the function}
}
\description{
Tune the lambda parameter in function acPCA
}
