% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acPCoA_prediction.R
\name{acPCoA_prediction}
\alias{acPCoA_prediction}
\title{Perform AC-PCoA on new data points for classification and prediction}
\usage{
acPCoA_prediction(
  training.dist,
  testing.dist,
  training.confounder,
  training.label,
  testing.label,
  nPC = 2,
  kernel = "linear",
  bandwidth = NULL,
  lambdas = seq(0, 5, 0.05),
  anov = T,
  perc = 0.05
)
}
\arguments{
\item{training.dist}{the n by n data distance matrix, where n is the number of training samples.}

\item{testing.dist}{the n by m data distance matrix, where n is the number of training samples and m is the number of testing samples.}

\item{training.confounder}{the n by q confounder matrix of training samples, where n is the number of samples, q is the number of confounding factors.}

\item{training.label}{Label of traning samples for classification. Optional.}

\item{testing.label}{Label of testing samples for classification. Optional.}

\item{nPC}{number of principal components to compute}

\item{kernel}{the kernel to use: "linear", "gaussian".}

\item{bandwidth}{bandwidth h for Gaussian kernel. Optional.}

\item{lambdas}{the tuning parameter, non-negative.}

\item{anov}{True or False. Whether the penalty term has the between groups sum of squares interpretation. Default is True.}

\item{perc}{the best lambda is defined to be the smallest lambda with R(lambda)<=perc (if anov=T), or R(lambda)<=perc*R(lambda=0) (if anov=F) in the nPC principal components.}
}
\value{
principal components and the projected data of training and testing samples
\item{v}{the principal components, p by nPC matrix}
\item{Xv_training}{the projected data of training samples, i.e. X times v}
\item{Xv_testing}{the projected data of testing samples}
\item{eigenvalueX}{eigenvalues for the PCs}
\item{varianceX}{variance explained by the PCs}
\item{PredictingAccuracy}{the accuracy of predicting labels of testing samples using Random Forest, if training and testing labels are provided.}
}
\description{
Perform AC-PCoA on new data points for classification and prediction
}
\examples{
\dontrun{
training.dist=data_mbqc_groupA$DistMat.BC[1:679,1:679]
testing.dist=data_mbqc_groupA$DistMat.BC[1:679,680:848]
training.confounder=data_mbqc_groupA$ConfounderMat[1:679,]
label=as.factor(data_mbqc_groupA$Specimen)
training.label=label[1:679]
testing.label=label[680:848]
result_prediction=acPCoA_prediction(training.dist,testing.dist,training.confounder,training.label,testing.label,nPC=2,kernel="linear",bandwidth=NULL,lambdas=seq(0, 5, 0.05),anov=T,perc=0.05)
}
}
