% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calkernel.R
\name{calkernel}
\alias{calkernel}
\title{Calculate the kernel matrix}
\usage{
calkernel(Y, kernel, bandwidth, scaleY = F)
}
\arguments{
\item{Y}{the n by q confounder matrix, where n is the number of samples, q is the number of confounding factors. Missing values in Y should be labeled as NA.}

\item{kernel}{the kernel to use: "linear", "gaussian".}

\item{bandwidth}{bandwidth h for Gaussian kernel. Optional.}

\item{scaleY}{scale the columns in Y to unit standard deviation. Default is False.}
}
\value{
The kernel matrix
\item{K}{the n by n kernel matrix for Y}
}
\description{
Calculate the kernel matrix
}
\examples{
Y <- data_tree$ConfounderMat
K1 <- calkernel(Y, kernel="linear") ##linear kernel
K2 <- calkernel(Y, kernel="gaussian", bandwidth=1) ##Gaussian kernel
}
