% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mbqc_groupA.R
\docType{data}
\name{data_mbqc_groupA}
\alias{data_mbqc_groupA}
\title{MBQC data, group A.}
\format{
A list with multiple elements, which are
\describe{
  \item{DistMat.EU}{the n by n distance matrix. n is the number of samples. Euclidean distance is applied.}
  \item{DistMat.BC}{the n by n distance matrix. n is the number of samples. Bray-Curtis distance is applied.}
  \item{ConfounderMat}{the n by q confounder matrix}
  \item{Specimen}{The specimen label of samples}
  \item{BioLab}{The bioinformatics labs which processed the samples}
  \item{BioID}{The sample ID from biological labs before sending to the bioinformatics labs.}
}
}
\usage{
data_mbqc_groupA
}
\description{
A subset of 848 samples from MBQC are included in the group A.
}
\keyword{datasets}
